Updated project to be able to push to PyPI
