+++ redfish_interop_validator/RedfishLogo.py
